+++ addons/pt/saft/invoices.go
@@ -26,7 +26,7 @@ var (
 )
 
 func validateInvoice(inv *bill.Invoice) error {
-	it := invoiceType(inv)
+	dt, _ := DocType(inv)
 
 	return validation.ValidateStruct(inv,
 		validation.Field(&inv.Tax,
@@ -34,11 +34,11 @@ func validateInvoice(inv *bill.Invoice) error {
 			validation.Skip,
 		),
 		validation.Field(&inv.Series,
-			validateSeriesFormat(it),
+			validateSeriesFormat(dt),
 			validation.Skip,
 		),
 		validation.Field(&inv.Code,
-			validateCodeFormat(inv.Series, it),
+			validateCodeFormat(inv.Series, dt),
 			validation.Skip,
 		),
 		validation.Field(&inv.Lines,
@@ -62,10 +62,29 @@ func validateTax(val any) error {
 
 	return validation.ValidateStruct(t,
 		validation.Field(&t.Ext,
-			tax.ExtensionsRequire(ExtKeyInvoiceType),
+			validation.By(validateTaxExt),
 			validation.Skip,
 		),
 	)
+}
+
+func validateTaxExt(val any) error {
+	ext, _ := val.(tax.Extensions)
+	if ext == nil {
+		ext = make(tax.Extensions) // Empty temporary map to return meaningful errors
+	}
+
+	msg := fmt.Sprintf("either `%s` or `%s` must be set", ExtKeyWorkType, ExtKeyInvoiceType)
+
+	if !ext.Has(ExtKeyWorkType) && !ext.Has(ExtKeyInvoiceType) {
+		return validation.NewError("invalid", msg)
+	}
+
+	if ext.Has(ExtKeyWorkType, ExtKeyInvoiceType) {
+		return validation.NewError("invalid", msg+", but not both")
+	}
+
+	return nil
 }
 
 // validateSeriesFormat validates the format of the series to meet the requirements of the
@@ -122,12 +141,4 @@ func validateCodeFormat(series cbc.Code, docType cbc.Code) validation.Rule {
 		}
 		return nil
 	})
-}
-
-func invoiceType(inv *bill.Invoice) cbc.Code {
-	if inv == nil || inv.Tax == nil || inv.Tax.Ext == nil {
-		return cbc.CodeEmpty
-	}
-
-	return inv.Tax.Ext[ExtKeyInvoiceType]
 }